Create Ideal-Money
Is there such a thing as 'perfect' money? Let's find out as we explore this question of what money is, can be, should be, and essentially must be, in order for a healthy economy to function. Get paid $EPIC Cash to learn about the future of blockchain-based currency in the "Ideal Money" series with your host and guide, Milton Freeman.